fix auto maxthreads

+++ src/service/ServiceSettings.cpp
@@ -13,7 +13,7 @@ QUrl const& ServiceSettings::url() const
     return m_url;
 }
 
-size_t ServiceSettings::maxImageConvertorThreads() const
+int ServiceSettings::maxImageConvertorThreads() const
 {
     return m_maxImageConvertorThreads;
 }
@@ -23,7 +23,7 @@ void ServiceSettings::setUrl(QUrl const& url)
     m_url = url;
 }
 
-void ServiceSettings::setMaxImageConvertorThreads(size_t maxImageConvertorThreads)
+void ServiceSettings::setMaxImageConvertorThreads(int maxImageConvertorThreads)
 {
     m_maxImageConvertorThreads = maxImageConvertorThreads;
 }
